+++ core-ui/src/state/activeNamespaceIdAtom.ts
@@ -2,7 +2,8 @@ import { atom, RecoilState } from 'recoil';
 
 type ActiveNamespaceIdState = string;
 
-const defaultValue = '';
+//empty value here would mean '[*]' - all namespaces
+export const defaultValue = '--no-namespace';
 
 export const activeNamespaceIdState: RecoilState<ActiveNamespaceIdState> = atom<
   ActiveNamespaceIdState
