+++ core-ui/src/state/navigation/navigationNodesSelector.ts
@@ -1,17 +1,20 @@
 import { RecoilValueReadOnly, selector } from 'recoil';
 import { isEmpty } from 'lodash';
 import { resourceListSelector } from '../resourceList/resourceListSelector';
-import { fetchPermissions, PermissionSet } from './fetchPermissions';
-import { activeNamespaceIdState } from '../activeNamespaceIdAtom';
+import {
+  activeNamespaceIdState,
+  defaultValue as defaultNamespaceName,
+} from '../activeNamespaceIdAtom';
 import { openapiPathIdListSelector } from '../openapi/openapiPathIdSelector';
 import { configFeaturesState } from '../configFeaturesAtom';
 import { filterExistingNodes } from './filters/filterExistingNodes';
 import { filterScopeNodes } from './filters/filterScopeNodes';
 import { filterNodesWithDisabledFeatures } from './filters/filterNodesWithDisabledFeatures';
 import { assignNodesToCategories } from './assignToCategories';
+import { filterPermittedNodes } from './filters/filterPermittedNodes';
+import { permissionSetsAtom } from '../permissionSetsAtom';
 import { Category } from './categories';
 import { NavNode } from '../types';
-import { filterPermittedNodes } from './filters/filterPermittedNodes';
 
 export const navigationNodesSelector: RecoilValueReadOnly<Category[]> = selector<
   Category[]
@@ -22,10 +25,20 @@ export const navigationNodesSelector: RecoilValueReadOnly<Category[]> = selector
     const activeNamespaceId = get(activeNamespaceIdState);
     const openapiPathIdList = get(openapiPathIdListSelector);
     const configFeatures = get(configFeaturesState);
+    const permissionSet = get(permissionSetsAtom);
 
-    if (!openapiPathIdList || !configFeatures || isEmpty(resourceList))
+    console.log(permissionSet);
+
+    const areDependenciesInitialized =
+      openapiPathIdList &&
+      configFeatures &&
+      !isEmpty(resourceList) &&
+      activeNamespaceId !== defaultNamespaceName &&
+      !isEmpty(permissionSet);
+
+    if (!areDependenciesInitialized) {
       return [];
-    const permissionSet = (await fetchPermissions(get)) as PermissionSet[];
+    }
 
     const scope = activeNamespaceId ? 'namespace' : 'cluster';
